+++ admb-lang/src/index.ts
@@ -1,0 +1,2 @@
+export { admb } from "./admb";
+export { CompletionProvider as AdmbCompletionProvider } from "./completion";